+++ Cpp_02/ex03/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: annabrag <annabrag@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 16:47:57 by annabrag          #+#    #+#             */
-/*   Updated: 2025/05/22 22:27:20 by annabrag         ###   ########.fr       */
+/*   Updated: 2025/05/26 20:39:41 by annabrag         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ Fixed::Fixed(const float nbr)
 	// std::cout << BOLD PINK "[Float constructor] " RESET << "called" << std::endl;
 	if (nbr <= -8388607.1 || nbr >= 8388607.1)
 		throw std::invalid_argument("float is out of range");
-	Fixed::setRawBits(roundf(nbr * (1 << _fracBits)));
+	Fixed::setRawBits(static_cast<int>(roundf(nbr * (1 << _fracBits))));
 }
 
 Fixed&	Fixed::operator=(const Fixed& toCopy)
